fix: 拆分代码
